Improved tree handling in Pi0 physics class.
A two stage physics analysis is now implemented in the example PPi0 class. A basic tree is created containing only the calculated Missing Mass and Energy, along with the TC, incident photon energy, theta, phi and time for timing cuts. Histograms can be produced from this tree (at the same time) without interfering with the tree. This is done so that the histogram binning can be changed on the fly (as well as missing mass cuts and prompt/random timing cuts). All controls are still hard coded into the init section, however this will change to be user controlled (either via command line flags, or through a config file).