Changes to Charged-Neutral conditions (Particle reconstruct)
Setting up basic outline for improved "charged particle marking". This is a seperate parameter to full charged particle identification (ie dE_E cuts) which looks at the energy deposition in the charged particle detectors to mark a particle as charged. Due to experimental problems with individual detectors, this must be done in a way which is possible to completely ignore the information from an individual detector.