Included TDatabasePDG from ROOT - CMakeLists updated temporarily
TDatabasePDG provides particle properties (read in from $ROOTSYS/etc/pdg_table.txt)
which provides the ability to get particle properties like mass, pdgcode, lifetime, etc.
GSort and GParticleRecon. have been updated to make use of this capability.
A library which is not normally compiled with the ROOT_LIBRARIES flag has been included
in CMakeLists by hand, but this linking will be changed to a better method in a
future commit.