Improved storage of sorts, makes use of TDatabasePDG
New storage of sorts allows for calls to the TDatabasePDG to gather PDG codes.
Cuts can be turned on my name (ex. proton) or by PDG (ex. 2212). An error in
the neutrality sort was also corrected (the neutral cut had an improper check
on the charge, fixed.) Multiple lines for the same particle type can also be
turned on (ex. 2+ gammas AND 5- gammas would accept events with
2, 3, 4, or 5 photons).