Changes to meson reconstruction
Rejection of Ek == 0 particles from meson reconstruction is included.
This is needed as the CentralApparatus class will create particle tracks
using the PID and MWPC if possible. In this case, the Ek is 0 as no
calorimeter information is available. If these particles are allowed
to be included in the meson reconstruction, they are just included
as an additional daughter (without actually changing the lorentz vector).
For this reason, they shouldn't be included.